Try new

+++ src/features/cart/cartSlice.js
@@ -1,43 +1,49 @@
-import { createSlice } from "@reduxjs/toolkit";
+import { createSlice } from '@reduxjs/toolkit';
 
 const initialState = {
   cart: [],
-  //   cart: [
-  //     {
-  //       pizzaId: 12,
-  //       name: "Mediterranean",
-  //       quantity: 2,
-  //       unitPrice: 16,
-  //       totalPrice: 32,
-  //     },
-  //   ],
+
+  // cart: [
+  //   {
+  //     pizzaId: 12,
+  //     name: 'Mediterranean',
+  //     quantity: 2,
+  //     unitPrice: 16,
+  //     totalPrice: 32,
+  //   },
+  // ],
 };
+
 const cartSlice = createSlice({
-  name: "cart",
+  name: 'cart',
   initialState,
   reducers: {
     addItem(state, action) {
+      // payload = newItem
       state.cart.push(action.payload);
     },
     deleteItem(state, action) {
-      state.cart = state.cart.filter((item) => item.pizzaId != action.payload);
+      // payload = pizzaId
+      state.cart = state.cart.filter((item) => item.pizzaId !== action.payload);
     },
     increaseItemQuantity(state, action) {
-      const item = state.cart.find((item) => item.pizzaId == action.payload);
+      // payload = pizzaId
+      const item = state.cart.find((item) => item.pizzaId === action.payload);
+
       item.quantity++;
       item.totalPrice = item.quantity * item.unitPrice;
     },
     decreaseItemQuantity(state, action) {
-      const item = state.cart.find((item) => item.pizzaId == action.payload);
+      // payload = pizzaId
+      const item = state.cart.find((item) => item.pizzaId === action.payload);
+
       item.quantity--;
       item.totalPrice = item.quantity * item.unitPrice;
-      if (item.quantity == 0) {
-        // state.cart = state.cart.filter((item) => item.pizzaId != action.payload);
-        cartSlice.caseReducers.deleteItem(state, action);
-      }
+
+      if (item.quantity === 0) cartSlice.caseReducers.deleteItem(state, action);
     },
     clearCart(state) {
-      state.cart = initialState.cart;
+      state.cart = [];
     },
   },
 });
@@ -49,15 +55,18 @@ export const {
   decreaseItemQuantity,
   clearCart,
 } = cartSlice.actions;
+
 export default cartSlice.reducer;
 
-export const getTotalCartPrice = (state) =>
-  state.cart.cart.reduce((acc, item) => acc + item.totalPrice, 0);
+export const getCart = (state) => state.cart.cart;
 
 export const getTotalCartQuantity = (state) =>
-  state.cart.cart.reduce((acc, item) => acc + item.quantity, 0);
+  state.cart.cart.reduce((sum, item) => sum + item.quantity, 0);
 
-export const getPizzaExistsInCart = (state, id) => {
-  const item = state.cart.cart.find((pizza) => pizza.pizzaId === id);
-  return item ? item.quantity : null
-};
+export const getTotalCartPrice = (state) =>
+  state.cart.cart.reduce((sum, item) => sum + item.totalPrice, 0);
+
+export const getCurrentQuantityById = (id) => (state) =>
+  state.cart.cart.find((item) => item.pizzaId === id)?.quantity ?? 0;
+
+// 'reselect'
